+++ ReleaseNotes.md
@@ -1,1 +1,2 @@
-* First release!
+* Removing Scatterer hard dependency
+* Using PartialDepthBuffer from Scatterer if possible